Factor out config hash.
